backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.23.1
# Use a newer pydantic patch release for compatibility with Python 3.12
pydantic==2.12.5
# Pin SQLAlchemy to a 2.0.x release compatible with the environment
sqlalchemy==2.0.45
psycopg2-binary==2.9.6 # 外部の管理ツール、ローカルの同期クライアントなどのために保留
asyncpg==0.31.0
pytest==9.0.2
# httpx pin was causing conflicts between litellm and google-genai.
# Allow a compatible range required by google-genai (>=0.28.1).
httpx>=0.28.1,<1.0

# Template rendering
Jinja2==3.1.2
litellm==1.80.11
google-genai==1.56.0
arq==0.26.3
aioredis==2.0.1
pytest-asyncio==1.3.0
types_pyyaml==6.0.12
pytest_mypy_plugins==3.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
itsdangerous==2.2.0
